Fixes for breakage in GIT import from SVN

+++ include/FXGenericTools.h
@@ -2175,7 +2175,7 @@ namespace TL
 			{
 				FXSTATIC_ASSERT(length<typelist>::value<=16, DynamicAt_Maximum_Exceeded);
 				typedef instance<NullType> nullinst;
-				if(idx>=length<typelist>::value)
+				if((int)idx>=length<typelist>::value)
 					return &nullinst::Do;
 				typedef typename apply<typelist, instance>::value instancedtypelist;
 				typedef typename replicate<16-length<typelist>::value, instance<NullType> >::value instancedtypelistend;
